extstore: prefill global page pool with extstore
the slab page mover algo would fill memory to evicting for a few seconds
before jumping to life and shoveling some pages into the global pool.

swore I was going to fix this post-release, but I had a moment of inspiration
after finding some code from another branch that did half the work. After a
bunch of stupid bugs it seems to work.

-o slab_automove_freeratio=0.N is now an option. This is *percentage of total
memory*, so don't set it too high.