limit crawls for metadumper
LRU crawler metadumper is used for getting snapshot-y looks at the LRU's.
Since there's no default limit, it'll get any new items added or bumped since
the roll started.

with this change it limits the number of items dumped to the number that
existed in that LRU when the roll was kicked off. You still end up with an
approximation, but not a terrible one:

- items bumped after the crawler passes them likely won't be revisited
- items bumped before the crawler passes them will likely be visited toward
  the end, or mixed with new items.
- deletes are somewhere in the middle.