extstore: fix ref leak when using binprot GATK
GATK returns a key but not the value. c->io_wraplist is only appended if the
value is to be returned, but c->item is skipped if it is an ITEM_HDR at all.

This now checks for the ITEM_HDR bit being set but also !value
which then reclaims the reference normally.

I knew doubling up the cleanup code made it a lot more complex, and hope to
flatten that to a single path. Also the TOUCH/GAT/GATK binprot code has no real
test coverage, nor mc-crusher entries. Should be worth fixing.