mget: autovivify (N)
toying around with where this goes in the code.

there's an option of autovivifying _before_ doing all of the return
tests, then simply switching VALUE to WIN in some cases.

else we get a truncated WIN line instead of the original VALUE line.

the former is probably what we want or else I have to figure CAS/etc.
committing this so I can fiddle/test without losing position.