metaget: error handling
getting a lot closer to liking this... goto is blegh but they can be
refactored out if refactored into a couple more functions.. and
generally it's just fine.

still a problem: the results of SET are being eaten by noreply when we
want to return text on CAS failure (EXISTS).

also on mdelete: responses should contain keys? I _think_ it's fine if
responses only contain keys sometimes, but will mull that over a bit.