restart: refactor into restart.c
pulls a lot of the shouldn't-be-in-slabs.c code out into restart.c,
managed via memcached.c

Code is mostly back to where it was before; it starts and seems to
recover data. The mess with tracking p->mem_requested was fixed in an
earlier commit, which simplifies the recovery process.

Still more refactoring needed.