restart: add missing msync at close time
To make sure that the backing pages actually make it to the backing
device prior to shutdown, munmap must be preceded by an msync call.
Additionally, since the metadata file is used as a flag to check if
the shutdown was graceful, the data needs to be sync'd prior to that
file being written.