Keep "last access" time up to date in SLRU mode
The last access time used to only update once per minute to avoid
excess bumping on hot items. However, with segmented mode if an item is
hit a lot it's simply poked in place.

Previous to this change we were calling extra functions and branches
for no real reason. Also when bumping within the WARM_LRU, we were
updating the last access time despite it being a shuffle. Also it was
skipping the bump if the access time was too recent, which is one hell
of a bug.