network: simplify client redispatching
redispatching from a side thread (extstore/crawl/etc) reused the new
connection dispatching CQ_item() structure, which is globally locked.
It can also fail to malloc sometimes, which would lead to a stranded
client.

This used to just close the connection, but I want to only close client
connections in its original thread. Using the same mechanism as the
idle connection kicker (just writing the FD to the pipe) removes a lock
bottleneck and the potential error, so we don't need to close the
socket anymore.