test fix: watcher.t flaky on travis CI
I cannot get this to fail on my quadcore desktop, an rpi3, or a minspec
digitalocean VM, despite waiting 30+ minutes with test looping. Yet
travis manages to do it.

Have a feeling this'll fix it. Worth a shot, anyway. There's no way to
programatically tell when all of the workers have gotten their logs
enabled.