Add BoringSSL support.
This won't matter in *nix since OpenSSL/LibreSSL is shipped by default(?) but in Windows, statically linking TLS library is preferred just like cURL and Chromium. Statically linking with BoringSSL generates at least 30% smaller executable than with OpenSSL.
BoringSSL is an OpenSSL derivative and is mostly source-compatible, for the subset of OpenSSL retained. SSL_in_before API used in memcached is not available in BoringSSL but this can be disabled. memcached is using this API for rejecting peer renegotiations which is already the default in BoringSSL.