net: carve response buffers from read buffers
balancing start arguments for memory limits on read vs response buffers
is hard. with this change there's a custom sub-allocator that cuts the
small response objects out of larger read objects.

it is somewhat complicated and has a few loops. the allocation path is
short circuited as much as possible.