portability fix for getsubopt
in the deferred IO patch I'd cribbed some code from another PR which
turned out to use a non-portable return state from getsubopt. On BSD
systems the subopts_value string is simply empty on a non-match.

Fixing this portably looks gross but I couldn't come up with a better
option... and frankly if we're going to pass options through DSO
submodules in the future this is better anyway.