idle_timeout: avoid long hangs during shutdown
the idle thread would sleep for up to the max timeout given before
checking for active connections again. the new shutdown code waits on
threads to be joined after a shutdown signal, which could end up
hanging by the idle_timeout.

this now instead uses a cond_timedwait call to allow a condition signal
to wake up the thread to exit.

this should be done for a few other threads that sleep without
conditions (lru_maintainer?)