[docker] Add user and expose the memcached port
These changes aim to improve our memcached devtools images to allow us
to reuse them during development stages.

These changes create a memcached user for all the images, switch to this
user by default once memcached have been configured for building.

That allow to run the memcached server directly when the containers are
started, else, without the user the run will fail as we try to execute
memcached as root user.

Also we should notice that the alpine image has been fixed by adding git
to the list of the requirements to install. Else the build will fails.