core: io_queue flow second attempt
probably squash into previous commit.

io->c->thead can change for orpahned IO's, so we had to directly add the
original worker thread as a reference.

also tried again to split callbacks onto the thread and off of the
connection for similar reasons; sometimes we just need the callbacks,
sometimes we need both.