proxy: simplified bad backend handling
Taking a simplified first attempt at having fast-failures for backends
that throw errors or time out; IO's will immediately be returned for
these backends until they become alive. A background retry is scheduled
until the backend works again.

There are a lot of ways to improve this, some noted in comments, and I
will likely need to move the connect() code anyway, so doing that as a
later change.