proxy: start logger integration
timing is a little tough; not a lot of places to stick the time start vs
time end for the log. gives a raw timing of microseconds.

also need to follow up with having a response inherit some details about
the request (type?) and translate the response type/code to something
useful... but starting with this.