proxy: start of io_uring integration
This commit changes:
 - add --enable-proxy-uring option, which looks for a compiled liburing
   in vendor/liburing
 - change backend drive_machine() to remove syscalls so it can be a
   common code path between libevent and uring
 - add basic liburing support for the backend read path. writes to
   backends still use normal syscalls and will be optimized later

The structure will need to change to meter out *sqe's more carefully,
but it should be impossible to trip the exhaustion condition so this is
fine for now.

The mcmc code did not make this easy, drive_machine() has turned into a
bit of a mess. I need to flatten that out to allow avoiding memmove's
and make the code clearer. Will come back to it after taking a break
from it...

Uring path is missing some critical code still:
 - backend disconnects
 - bad backend handling/retry
 - timeouts don't exist at all

So it should work for testing in good conditions and I'll add the error
handling during the stabilization phase.