Fix for #837 - tests fail on OS X
Always at least try to setrlimit to what we want it to be, but in debug
mode ignore if it fails. This isn't the greatest workaround but will
ensure the next release works under both valgrind and OS X. A better way
to detect valgrind (importing their huge .h file I guess?) would mean we
can skip setrlimit in that specific case only.