meta: fix CAS ('c') return values
This should now align with the binary protocol. It might make sense for
this to return the _current_ CAS value of the underlying item in the
cases of EX... binprot does not currently do this, and the storage code
is shared between binary and text so I can't change them independently.