core: fix use-after-free for text multigets
Reported in #849 - this fixes copying a read buffer after freeing the
original read buffer.

This didn't matter for years since the cache code didn't touch the
buffer, but recently it can reuse the first 8 bytes as a pointer to the
internal freelist. Thus in some situations where large reads happen the
command can get corrupted, returning an unhelpful "ERROR" to the end
user.