proxy: fix bug when enabled but set not overridden
if we're falling back to ascii protocol handling and that command relies
on complete_nread_ascii, we were still landing in complete_nread_proxy.

this is an incomplete fix as we should explicitly handle for a "bug" of
having a nil coroutine here. Requires some restructuring so that's for
v2.