proxy: counter for backend failures
think this ony counts failues related to the connect routine, and when a
backend is considered dead. Looks like I was partway through deciding
how to change that.

I'm sure it's possible for something to fail in such a way that it
connects, then immediately fails, which escapes this routine. Separate
fix for that though.