proxy: 'stats proxy' crashes on invocation
I noticed running 'stats proxy' caused a crash. On observing it under
GDB, it crashed while attempting to access the worker thread level
proxy_user_stats in process_proxy_stats().
This showed that the thread level 'proxy_user_stats' wasnt properly
initialized. On checking it's initialization code in
proxy_worker_reload(), a typo was noticed that caused the bug.
This patch fixes it.