proxy: fix crash on `stats proxy` sans user stats
Fix is a little subtle: there's a race condition when a user stat is
added to the global context but the worker threads aren't updated yet.
So we check that the context in the user thread has enough stats slots
to iterate, and that the user thread has stats at all.