proxy: add r:has_flag(), fix r:token() length
Add mcp.request function for quick checking if a flag exists in a
request string.

Also updates internal code for checking the length of a token to use the
endcap properly, and uses that for the r:token(n) requets as well, which
fixes a subtle bug of the token length being too long.