proxy: fix subtle parser bug for some cmds
in the request parser the new endcap token was always the end of the
line, but it should be the start of the next token. For some use cases
the scanning stops early, or if we have too many tokens it would make
the final token look like the rest of the line.