Race leads to deadlock during shutdown (sigterm/sigusr1)
We've noticed in some rare cases, that sending a sigterm to a memcached process
could lead to a race condition which leads to a deadlock of the process.

The worker threads are stuck on 'worker_hang_lock' because the shutdown
process hasn't completed yet.

The shudown process hasn't completed since the logger thread signals
'logger_stack_cond' without locking its corresponding 'logger_stack_lock'.
This could cause the signal to get lost, causing the logger thread to
wait indefinitely on 'logger_stack_cond' in logger_thread().