proxy: io_uring work and simplify write flush
expands on the io_uring port for the IO thread, moving write calls
behind the ring.

unfortunately this does not result in any performance uplift. I also
tried moving the eventfd related writes (which was backed out later).
The code still needs to be updated to generate SQE's outside of the CQE
loop (see comments above the uring event loop).

It is still generally buggy and not to be used, but keeping it up to
date every once in a while will let us use it if it ever does become
performant enough...

This also makes use of a prior optimization to simplify the write flush
code a little. It no longer needs to track how many bytes were intended
to write to know if it should continue flushing later.