proxy: iterate modified request handling
have a bug where updating a token and then requesting it again returns
the previous token. Also have other branches which require use of the
flattened request.

Removes the extra allocation space for lua request objects as we're not
flattening into the end of the memory.

I was originally doing this using a lot of lua but just copying the
string a few times has some better properties:
1) should actually be faster with less lua + fewer allocations
2) can be optimized to do minimal copying (avoid keys, append new flags,
   etc)