proxy: fix "missingend" error on reading responses
If the backend handler reads an incomplete response from the network, it
changes state to wait for more data. The want_read state was considering
the data completed if "data read" was bigger than "value length", but it
should have been "value + result line".

This means if the response buffer landed in a bullseye where it has read
more than the size of the value but less than the total size of the
request (typically a span of 200 bytes or less), it would consider the
request complete and look for the END\r\n marker.

This change has been... here forever.