crawler: add "lru_crawler mgdump" command
The "metadump" command was designed primarily for doing analysis on
what's in cache, but it's also used for pulling the data out for various
reasons.

The string format is a bit onerous: key=value (for futureproofing) and
URI encoded keys (which may or may not be binary internally)

This adds a command "mgdump", which dumps keys in the format:

"mg key\r\nmg key2\r\n"

if a key is binary encoded, it uses the meta binary encoding scheme of
base64-ing keys and appends a "b" flag:

"mg 44OG44K544OI b\r\n"

when the dump is complete it prints an "EN\r\n"

clients wishing to stream or fetch data can take the mg commands, strip
the \r\n, append any flags they care about, then send the command back
to the server to fetch the full key data.

This seems to use 30-40% less CPU time on the server for the same key
dumps.