proxy: fix data corruption bug
Bug introduced in 6c80728: use after free for response buffer while
under concurrency.

The await code has a different method of wrapping up a lua coroutine
than a standard response, so it was not managing the lifecycle of the
response object properly, causing data buffers to be reused before being
written back to the client.

This fix separates the accounting of memory from the freeing of the
buffer, so there is no more race.

Further restructuring is needed to both make this less bug prone and
make memory accounting be lock step with the memory freeing.