proxy: fix flaky test in proxyconfig.t
The connect timeout won't fire when blocking a backend from connecting
in these tests; it will connect, send a version command to validate,
then time out on read.

With the read timeout set to 0.1 it would sometimes fail before the
restart finished, clogging log lines and causing test failures.

Now we wait for the watcher and remove a sleep, with a longer read
timeout.