proxy: use connect timeout during retries
The connect timeout is supposed to be applied to TCP connect() calls.
The "retry timeout" is a poorly named variable for how long to wait
before retrying a backend that is marked as bad.

The retry timeout was accidentally being applied in cases where
connect() was being called when retrying broken connections.

This now uses the appropriate timeouts.