extstore: fix data bugs on high overwrite key
Scenario required:

- likely item_age=2 (or some low value), so cache memory is otherwise
  not used to hold data very long.
- one, or a few keys, are frequently updated and re-written to disk.
- the same key must be written to the same data page multiple times
- at some point extstore compaction runs

There are two faults to this bug:

1) We may accidentally "rescue" an older version of the frequently
   overwritten key. More likely the data will go missing.
2) More likely, extstore will log too many "deletes" against an active
   disk page. It will then think the page is empty when it isn't and
   invalidate it.
   This results in other random items disappearing when they shouldn't.

The fix is simple in retrospect: when validating an in-memory header
against data on disk, we check that the page and page version are the
same, but it was not checking that the *offset on disk* is correct. Now
it checks all three identifiers which fixes the bug.