proxy: io return optimization
Optimizes the return path of async IO objects.

There is a secondary object ("connection queue object") which was being
allocated to hold the IO object while it is being returned from an async
process into the worker thread.

This is unnecessary: the IO object can manage itself, which avoids an
extra cache allocation/free and some repeated mutex locks for every IO
object returned.

Yields a varying degree of performance improvement. Heavy usage of
extstore or proxy await objects can be as much as 10%+ improvement.