proxy: mcp.backend({ down = true }) option
Starts backend in a "bad" state, so any attempt to access it will result
in "SERVER_ERROR backend failure".

The only times a backend's bad flag can be removed is on a successful
connection attempt. So all we need to do is avoid fully setting up or
tearing down the backend and starting it marked bad.