extstore: fix item loss during page defrag
If an item that is in the process of being moved to a new disk page
during defragmentation, is also being accessed by a client connection,
we end up losing the item. This is because we cannot update the item
header unless we have exclusive ownership over the item.

This is an extremely rare scenario. You must:
- Be running with item_age=0 or a very low number
- Constantly causing defragmentation
- Get very unlucky during defragmentation.

With this patch we will attempt to re-allocate the item header if we
cannot overwrite it implicitly: this is the same process as happens with
arithmetic commands in do_add_delta().